+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/service/Schema.java
@@ -17,7 +17,6 @@
 package com.exonum.binding.core.service;
 
 import com.exonum.binding.core.blockchain.Block;
-import com.exonum.binding.core.blockchain.BlockchainData;
 import com.exonum.binding.core.transaction.Transaction;
 
 /**
@@ -33,8 +32,8 @@ import com.exonum.binding.core.transaction.Transaction;
  * <p>Exonum starts aggregating a service collection state hash once it is <em>initialized</em>:
  * created for the first time with a read-write
  * {@link com.exonum.binding.core.blockchain.BlockchainData} (e.g., in a
- * {@linkplain Service#initialize(BlockchainData, Configuration) service constructor},
- * or in a {@linkplain Transaction}).
+ * {@linkplain Service#initialize(com.exonum.binding.core.transaction.TransactionContext,
+ * Configuration) service constructor}, or in a {@linkplain Transaction}).
  *
  * <p>Please note that if the service does not use any Merkelized collections,
  * the framework will not be able to verify that its transactions cause the same